requirements.txt
curl_cffi>=0.7.0
beautifulsoup4>=4.12.0
easyocr>=1.7.1
pillow>=10.0.0
numpy>=1.26.0
openai>=1.12.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.9
